- Converted floats to doubles - added Reference Getters returning the Ref to the object instead of the pointer

+++ FusionLib/include/Utils.hpp
@@ -8,9 +8,9 @@ class Point2D{
 public:
 
     Point2D():_data(0),_position(0,0,1){}
-    Point2D(int x,int y, float data):_data(data),_position(x,y,1),xx(x),yy(y){}
-    Eigen::Vector3f _position;
-    float _data;
+    Point2D(int x,int y, double data):_data(data),_position(x,y,1),xx(x),yy(y){}
+    Eigen::Vector3d _position;
+    double _data;
     int xx,yy;
 };
